Split engine_animator_unittests
