+++ ci/licenses_golden/excluded_files
@@ -252,6 +252,7 @@
 ../../../flutter/shell/common/context_options_unittests.cc
 ../../../flutter/shell/common/dl_op_spy_unittests.cc
 ../../../flutter/shell/common/engine_unittests.cc
+../../../flutter/shell/common/engine_animator_unittests.cc
 ../../../flutter/shell/common/fixtures
 ../../../flutter/shell/common/input_events_unittests.cc
 ../../../flutter/shell/common/persistent_cache_unittests.cc
