+++ shell/common/animator.cc
@@ -133,9 +133,14 @@ void Animator::EndFrame() {
         frame_timings_recorder_->GetVsyncTargetTime());
 
     // Commit the pending continuation.
-    PipelineProduceResult result =
-        producer_continuation_.Complete(std::make_unique<FrameItem>(
-            std::move(layer_trees_tasks_), std::move(frame_timings_recorder_)));
+    std::vector<std::unique_ptr<LayerTreeTask>> layer_tree_task_list;
+    for (auto& [view_id, layer_tree_task] : layer_trees_tasks_) {
+      layer_tree_task_list.push_back(std::move(layer_tree_task));
+    }
+    layer_trees_tasks_.clear();
+    PipelineProduceResult result = producer_continuation_.Complete(
+        std::make_unique<FrameItem>(std::move(layer_tree_task_list),
+                                    std::move(frame_timings_recorder_)));
 
     if (!result.success) {
       FML_DLOG(INFO) << "Failed to commit to the pipeline";
@@ -197,8 +202,11 @@ void Animator::Render(int64_t view_id,
                                 "Animator::Render", /*flow_id_count=*/0,
                                 /*flow_ids=*/nullptr);
 
-  layer_trees_tasks_.push_back(std::make_unique<LayerTreeTask>(
-      view_id, std::move(layer_tree), device_pixel_ratio));
+  // Only inserts if the view ID has not been rendered before, ignoring
+  // duplicate Render calls.
+  layer_trees_tasks_.try_emplace(
+      view_id, std::make_unique<LayerTreeTask>(view_id, std::move(layer_tree),
+                                               device_pixel_ratio));
 }
 
 const std::weak_ptr<VsyncWaiter> Animator::GetVsyncWaiter() const {
